Add files via upload
This program presents a menu with options for performing skin tests for hazardous materials, non-hazardous materials, and foreign products that touch the skin. Users can choose an option and receive guidance on how to perform the respective skin test. The program loops until the user chooses to exit.

Users are encouraged to consult healthcare professionals for proper procedures when dealing with hazardous materials. For non-hazardous materials and foreign products, users are reminded to ensure the safety of the materials for skin contact before use.